Add "strict" parameter to ForEachLine().
Add "strict" parameter to ForEachLine() to pick between
jrutil.ReadLine() which works with Mac EOLs but is a good deal slower
than bufio.ReadString() which does not work with Mac EOLs.  Not that
even though jrutil.ReadLine() is a good deal slower than
bufio.ReadString() it still runs about as fast as unix2dos.